Add StateName type and related methods; enhance FSM state validation and error handling
